rework add user logic, move csv methods to their own module
